+++ src/transactions/transactionBuilder/transactionBuilder.ts
@@ -27,6 +27,7 @@ import { normalizeBundle } from "../../utils/normalizeBundle";
 import {
   AccountAuthenticator,
   AccountAuthenticatorEd25519,
+  AccountAuthenticatorNoAccountAuthenticator,
   AccountAuthenticatorSingleKey,
 } from "../authenticator/account";
 import {
@@ -394,18 +395,11 @@ export function generateSignedTransactionForSimulation(args: InputSimulateTransa
       transaction.secondarySignerAddresses ?? [],
       transaction.feePayerAddress,
     );
-    let secondaryAccountAuthenticators: Array<AccountAuthenticator> = [];
-    if (secondarySignersPublicKeys) {
-      secondaryAccountAuthenticators = secondarySignersPublicKeys.map((publicKey) =>
-        getAuthenticatorForSimulation(publicKey),
-      );
-    }
-    if (!feePayerPublicKey) {
-      throw new Error(
-        "Must provide a feePayerPublicKey argument to generate a signed fee payer transaction for simulation",
-      );
-    }
-    const feePayerAuthenticator = getAuthenticatorForSimulation(feePayerPublicKey);
+    const secondaryAccountAuthenticators: Array<AccountAuthenticator> = secondaryAccountAuthenticatorsForSimulation(
+      secondarySignersPublicKeys,
+      transaction.secondarySignerAddresses?.length ?? 0,
+    );
+    const feePayerAuthenticator = getAuthenticatorForSimulation(feePayerPublicKey!);
 
     const transactionAuthenticator = new TransactionAuthenticatorFeePayer(
       accountAuthenticator,
@@ -426,16 +420,9 @@ export function generateSignedTransactionForSimulation(args: InputSimulateTransa
       transaction.secondarySignerAddresses,
     );
 
-    let secondaryAccountAuthenticators: Array<AccountAuthenticator> = [];
-
-    if (!secondarySignersPublicKeys) {
-      throw new Error(
-        "Must provide a secondarySignersPublicKeys argument to generate a signed multi agent transaction for simulation",
-      );
-    }
-
-    secondaryAccountAuthenticators = secondarySignersPublicKeys.map((publicKey) =>
-      getAuthenticatorForSimulation(publicKey),
+    const secondaryAccountAuthenticators: Array<AccountAuthenticator> = secondaryAccountAuthenticatorsForSimulation(
+      secondarySignersPublicKeys,
+      transaction.secondarySignerAddresses.length,
     );
 
     const transactionAuthenticator = new TransactionAuthenticatorMultiAgent(
@@ -456,13 +443,28 @@ export function generateSignedTransactionForSimulation(args: InputSimulateTransa
     );
   } else if (accountAuthenticator instanceof AccountAuthenticatorSingleKey) {
     transactionAuthenticator = new TransactionAuthenticatorSingleSender(accountAuthenticator);
+  } else if (accountAuthenticator instanceof AccountAuthenticatorNoAccountAuthenticator) {
+    transactionAuthenticator = new TransactionAuthenticatorSingleSender(accountAuthenticator);
   } else {
     throw new Error("Invalid public key");
   }
   return new SignedTransaction(transaction.rawTransaction, transactionAuthenticator).bcsToBytes();
 }
 
-export function getAuthenticatorForSimulation(publicKey: PublicKey) {
+function secondaryAccountAuthenticatorsForSimulation(
+  secondarySignersPublicKeys: Array<PublicKey | undefined> | undefined,
+  numSecondarySigners: number,
+): Array<AccountAuthenticatorNoAccountAuthenticator | AccountAuthenticatorEd25519 | AccountAuthenticatorSingleKey> {
+  if (secondarySignersPublicKeys) {
+    return secondarySignersPublicKeys.map((publicKey) => getAuthenticatorForSimulation(publicKey));
+  }
+  return new Array(numSecondarySigners).fill(undefined).map((publicKey) => getAuthenticatorForSimulation(publicKey));
+}
+
+export function getAuthenticatorForSimulation(publicKey?: PublicKey) {
+  if (!publicKey) {
+    return new AccountAuthenticatorNoAccountAuthenticator();
+  }
   // Wrap the public key types below with AnyPublicKey as they are only support through single sender.
   // Learn more about AnyPublicKey here - https://github.com/aptos-foundation/AIPs/blob/main/aips/aip-55.md
   const convertToAnyPublicKey =