+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to the Aptos TypeScript SDK will be captured in this file. T
 
 - Add the `isPrimitive` function to `TypeTag`.
 - Add `showStdout` optional property to `Move` and `LocalNode` classes to control the output of the CLI commands
+- Add support for MultiKey's in transaction simulations
 
 # 1.29.1 (2024-10-09)